Add Karaf build profile
Project local Karaf distros are handy for devs to test their work
however is unneeded by autorelease builds and should not be released
as part of the Simultanious Release. Add a profile that is active by
default so that default behaviour is unchanged however allows the
autorelease project to disable building this module.

Issue: RELENG-86
Change-Id: I4cf58c195a9a3a94ed2aa43b20bc28ed04719c89
Signed-off-by: Thanh Ha <thanh.ha@linuxfoundation.org>